26.01
Podział klas na pliki, dodanie dodatkowych zabezpieczeń podczas wpisywania danych do rejestracji
