+++ Frame.cs
@@ -1,0 +1,89 @@
+﻿using System;
+
+
+namespace Przychodnia_Gdynia
+{
+    class Frame
+    {
+        public static void Przychodnia()
+        {
+            System.Console.WriteLine(@"██████╗ ██████╗ ███████╗██╗   ██╗ ██████╗██╗  ██╗ ██████╗ ██████╗ ███╗   ██╗██╗ █████╗ 
+██╔══██╗██╔══██╗╚══███╔╝╚██╗ ██╔╝██╔════╝██║  ██║██╔═══██╗██╔══██╗████╗  ██║██║██╔══██╗
+██████╔╝██████╔╝  ███╔╝  ╚████╔╝ ██║     ███████║██║   ██║██║  ██║██╔██╗ ██║██║███████║
+██╔═══╝ ██╔══██╗ ███╔╝    ╚██╔╝  ██║     ██╔══██║██║   ██║██║  ██║██║╚██╗██║██║██╔══██║
+██║     ██║  ██║███████╗   ██║   ╚██████╗██║  ██║╚██████╔╝██████╔╝██║ ╚████║██║██║  ██║
+╚═╝     ╚═╝  ╚═╝╚══════╝   ╚═╝    ╚═════╝╚═╝  ╚═╝ ╚═════╝ ╚═════╝ ╚═╝  ╚═══╝╚═╝╚═╝  ╚═╝
+                                                                                       ");
+        }
+        public static void Panel_Uzytkownika()
+        {
+            System.Console.WriteLine(@"██████╗  █████╗ ███╗   ██╗███████╗██╗         ██╗   ██╗███████╗██╗   ██╗████████╗██╗  ██╗ ██████╗ ██╗    ██╗███╗   ██╗██╗██╗  ██╗ █████╗ 
+██╔══██╗██╔══██╗████╗  ██║██╔════╝██║         ██║   ██║╚══███╔╝╚██╗ ██╔╝╚══██╔══╝██║ ██╔╝██╔═══██╗██║    ██║████╗  ██║██║██║ ██╔╝██╔══██╗
+██████╔╝███████║██╔██╗ ██║█████╗  ██║         ██║   ██║  ███╔╝  ╚████╔╝    ██║   █████╔╝ ██║   ██║██║ █╗ ██║██╔██╗ ██║██║█████╔╝ ███████║
+██╔═══╝ ██╔══██║██║╚██╗██║██╔══╝  ██║         ██║   ██║ ███╔╝    ╚██╔╝     ██║   ██╔═██╗ ██║   ██║██║███╗██║██║╚██╗██║██║██╔═██╗ ██╔══██║
+██║     ██║  ██║██║ ╚████║███████╗███████╗    ╚██████╔╝███████╗   ██║      ██║   ██║  ██╗╚██████╔╝╚███╔███╔╝██║ ╚████║██║██║  ██╗██║  ██║
+╚═╝     ╚═╝  ╚═╝╚═╝  ╚═══╝╚══════╝╚══════╝     ╚═════╝ ╚══════╝   ╚═╝      ╚═╝   ╚═╝  ╚═╝ ╚═════╝  ╚══╝╚══╝ ╚═╝  ╚═══╝╚═╝╚═╝  ╚═╝╚═╝  ╚═╝");
+        }
+        public static void Aktualnosci()
+        {
+            Console.WriteLine(@" █████╗ ██╗  ██╗████████╗██╗   ██╗ █████╗ ██╗     ███╗   ██╗ ██████╗ ███████╗ ██████╗██╗
+██╔══██╗██║ ██╔╝╚══██╔══╝██║   ██║██╔══██╗██║     ████╗  ██║██╔═══██╗██╔════╝██╔════╝██║
+███████║█████╔╝    ██║   ██║   ██║███████║██║     ██╔██╗ ██║██║   ██║███████╗██║     ██║
+██╔══██║██╔═██╗    ██║   ██║   ██║██╔══██║██║     ██║╚██╗██║██║   ██║╚════██║██║     ██║
+██║  ██║██║  ██╗   ██║   ╚██████╔╝██║  ██║███████╗██║ ╚████║╚██████╔╝███████║╚██████╗██║
+╚═╝  ╚═╝╚═╝  ╚═╝   ╚═╝    ╚═════╝ ╚═╝  ╚═╝╚══════╝╚═╝  ╚═══╝ ╚═════╝ ╚══════╝ ╚═════╝╚═╝");
+        }
+        public static void Rodo()
+        {
+            Console.WriteLine(@"██████╗  ██████╗ ██████╗  ██████╗ 
+██╔══██╗██╔═══██╗██╔══██╗██╔═══██╗
+██████╔╝██║   ██║██║  ██║██║   ██║
+██╔══██╗██║   ██║██║  ██║██║   ██║
+██║  ██║╚██████╔╝██████╔╝╚██████╔╝
+╚═╝  ╚═╝ ╚═════╝ ╚═════╝  ╚═════╝ ");
+            AdditionalFunctions.TimerDot(3);
+        }
+        public static void Wyjscie()
+        {
+            Console.Clear();
+            AdditionalFunctions.TimerDot(3);
+            Console.WriteLine(@"██████╗  ██████╗     ███████╗ ██████╗ ██████╗  █████╗  ██████╗███████╗███████╗███╗   ██╗██╗ █████╗ ██╗
+██╔══██╗██╔═══██╗    ╚══███╔╝██╔═══██╗██╔══██╗██╔══██╗██╔════╝╚══███╔╝██╔════╝████╗  ██║██║██╔══██╗██║
+██║  ██║██║   ██║      ███╔╝ ██║   ██║██████╔╝███████║██║       ███╔╝ █████╗  ██╔██╗ ██║██║███████║██║
+██║  ██║██║   ██║     ███╔╝  ██║   ██║██╔══██╗██╔══██║██║      ███╔╝  ██╔══╝  ██║╚██╗██║██║██╔══██║╚═╝
+██████╔╝╚██████╔╝    ███████╗╚██████╔╝██████╔╝██║  ██║╚██████╗███████╗███████╗██║ ╚████║██║██║  ██║██╗
+╚═════╝  ╚═════╝     ╚══════╝ ╚═════╝ ╚═════╝ ╚═╝  ╚═╝ ╚═════╝╚══════╝╚══════╝╚═╝  ╚═══╝╚═╝╚═╝  ╚═╝╚═╝
+                                                                                                      ");
+        }
+        public static void LogRej()
+        {
+            Console.Clear();
+            Console.WriteLine(@"██╗      ██████╗  ██████╗  ██████╗ ██╗    ██╗ █████╗ ███╗   ██╗██╗███████╗    ██╗██████╗ ███████╗     ██╗███████╗███████╗████████╗██████╗  █████╗  ██████╗     ██╗ █████╗ 
+██║     ██╔═══██╗██╔════╝ ██╔═══██╗██║    ██║██╔══██╗████╗  ██║██║██╔════╝   ██╔╝██╔══██╗██╔════╝     ██║██╔════╝██╔════╝╚══██╔══╝██╔══██╗██╔══██╗██╔════╝     ██║██╔══██╗
+██║     ██║   ██║██║  ███╗██║   ██║██║ █╗ ██║███████║██╔██╗ ██║██║█████╗    ██╔╝ ██████╔╝█████╗       ██║█████╗  ███████╗   ██║   ██████╔╝███████║██║          ██║███████║
+██║     ██║   ██║██║   ██║██║   ██║██║███╗██║██╔══██║██║╚██╗██║██║██╔══╝   ██╔╝  ██╔══██╗██╔══╝  ██   ██║██╔══╝  ╚════██║   ██║   ██╔══██╗██╔══██║██║     ██   ██║██╔══██║
+███████╗╚██████╔╝╚██████╔╝╚██████╔╝╚███╔███╔╝██║  ██║██║ ╚████║██║███████╗██╔╝   ██║  ██║███████╗╚█████╔╝███████╗███████║   ██║   ██║  ██║██║  ██║╚██████╗╚█████╔╝██║  ██║
+╚══════╝ ╚═════╝  ╚═════╝  ╚═════╝  ╚══╝╚══╝ ╚═╝  ╚═╝╚═╝  ╚═══╝╚═╝╚══════╝╚═╝    ╚═╝  ╚═╝╚══════╝ ╚════╝ ╚══════╝╚══════╝   ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝ ╚═════╝ ╚════╝ ╚═╝  ╚═╝");
+        }
+        public static void Logowanie()
+        {
+            Console.Clear();
+            Console.WriteLine(@"██╗      ██████╗  ██████╗  ██████╗ ██╗    ██╗ █████╗ ███╗   ██╗██╗███████╗
+██║     ██╔═══██╗██╔════╝ ██╔═══██╗██║    ██║██╔══██╗████╗  ██║██║██╔════╝
+██║     ██║   ██║██║  ███╗██║   ██║██║ █╗ ██║███████║██╔██╗ ██║██║█████╗  
+██║     ██║   ██║██║   ██║██║   ██║██║███╗██║██╔══██║██║╚██╗██║██║██╔══╝  
+███████╗╚██████╔╝╚██████╔╝╚██████╔╝╚███╔███╔╝██║  ██║██║ ╚████║██║███████╗
+╚══════╝ ╚═════╝  ╚═════╝  ╚═════╝  ╚══╝╚══╝ ╚═╝  ╚═╝╚═╝  ╚═══╝╚═╝╚══════╝");
+        }    
+        public static void Rejestracja()
+        {
+            Console.Clear();
+            Console.WriteLine(@"██████╗ ███████╗     ██╗███████╗███████╗████████╗██████╗  █████╗  ██████╗     ██╗ █████╗ 
+██╔══██╗██╔════╝     ██║██╔════╝██╔════╝╚══██╔══╝██╔══██╗██╔══██╗██╔════╝     ██║██╔══██╗
+██████╔╝█████╗       ██║█████╗  ███████╗   ██║   ██████╔╝███████║██║          ██║███████║
+██╔══██╗██╔══╝  ██   ██║██╔══╝  ╚════██║   ██║   ██╔══██╗██╔══██║██║     ██   ██║██╔══██║
+██║  ██║███████╗╚█████╔╝███████╗███████║   ██║   ██║  ██║██║  ██║╚██████╗╚█████╔╝██║  ██║
+╚═╝  ╚═╝╚══════╝ ╚════╝ ╚══════╝╚══════╝   ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝ ╚═════╝ ╚════╝ ╚═╝  ╚═╝");
+        }
+    }
+}
