+++ Menu_Opcje.cs
@@ -1,0 +1,77 @@
+﻿using System;
+using static System.Console;
+
+
+namespace Przychodnia_Gdynia
+{
+    class Menu_Opcje
+    {
+        public static void USER_PANEL()
+        {
+            //Uzytkownik Franek = new Uzytkownik(false, true);
+            //Console.Clear();
+            //Frame.Panel_Uzytkownika();
+            //if(Franek.GetZalogowany() == true) Console.WriteLine("Status Uzytkownika: zalogowany");
+            //else WriteLine("Status Uzytkownika: niezalogowany");
+        }
+        public static void AKTUALNOSCI()
+        {
+            Console.Clear();
+            Frame.Aktualnosci();
+        }
+        public static void RODO()
+        {
+            Console.Clear();
+            Frame.Rodo();
+            System.Console.WriteLine();
+            System.Console.WriteLine();
+            Console.WriteLine("OCHRONA DANYCH OSOBOWYCH");
+            Console.WriteLine();
+            Console.WriteLine("1. Administratorem danych osobowych jest Przychodnia w Gdyni Sp. z o.o. 81-000   Gdynia ul. Gdyńska 1x.");
+            Console.WriteLine();
+            Console.WriteLine("2. Celem przetwarzania danych jest świadczenie przez Administratora usług medycznych.");
+            Console.WriteLine();
+            Console.WriteLine("3. Podstawą prawną przetwarzania jest art. 6 ust.1 lit. c) oraz art. 9 ust. 2 lit.h) RODO oraz przepisy polskich ustaw z obszaru prawa medycznego, pozostających w związku z celami zdrowotnymi przetwarzania, w szczególności ustawa o prawach pacjenta i Rzeczniku Praw pacjenta.");
+            Console.WriteLine();
+            Console.WriteLine("4. Administrator przekazuje dane osobowe pacjentów wyłącznie podmiotom uprawnionym na podstawie przepisów prawa lub podmiotów, z którymi zawarł umowę powierzenia przetwarzania danych osobowych.");
+            Console.WriteLine();
+            Console.WriteLine("5. Dokumentacja medyczna przechowywana jest przez okres 20 lat, licząc do końca roku kalendarzowego, w którym dokonano ostatniego wpisu , z wyjątkiem:");
+            Console.WriteLine();
+            Console.WriteLine("a) dokumentacji medycznej w przypadku zgonu pacjenta na skutek uszkodzenia ciała lub zatrucia, która jest przechowywana przez okres 30 lat, licząc od końca roku kalendarzowego, w którym nastąpił zgon;");
+            Console.WriteLine();
+            Console.WriteLine("b) dokumentacji medycznej zawierającej dane niezbędne do monitorowania losów krwi i jej składników, która jest przechowywana przez okres 30lat, licząc do końca roku kalendarzowego, w którym dokonano ostatniego wpis;");
+            Console.WriteLine();
+            Console.WriteLine("c) zdjęć rentgenowskich przechowywanych poza dokumentacją medyczną pacjenta,które są przechowywane przez okres 10 lat, licząc do końca roku kalendarzowego, w którym wykonano zdjęcie;");
+            Console.WriteLine();
+            Console.WriteLine("d) skierowań na badania lub zleceń lekarza, które są przechowywane przez okres:");
+            Console.WriteLine();
+            Console.WriteLine("    - 5 lat, licząc od końca roku kalendarzowego, w którym udzielono świadczenia zdrowotnego będącego przedmiotem skierowania lub zlecenia lekarza,");
+            Console.WriteLine();
+            Console.WriteLine("    - 2 lat, licząc od końca roku kalendarzowego, w którym wystawiono skierowanie - w przypadku gdy świadczenie zdrowotne nie zostało udzielone z powodu niezgłoszenia się pacjenta w ustalonym terminie, chyba że pacjent odebrał skierowanie;");
+            Console.WriteLine();
+            Console.WriteLine("    - dokumentacji medycznej dotyczącej dzieci do ukończenia 2 roku życia, która jest przechowywana przez okres 22 lat.");
+            Console.WriteLine();
+            Console.WriteLine("6. Osobie, której dane dotyczą przysługuje prawo dostępu do danych, ich sprostowania i uzupełnienia.");
+            Console.WriteLine();
+            Console.WriteLine("a. prawo dostępu do danych ( art. 15 RODO )");
+            Console.WriteLine();
+            Console.WriteLine("b. prawo do sprostowania i uzupełniania danych osobowych ( art. 16 RODO )");
+            Console.WriteLine();
+            Console.WriteLine("c. prawo do usunięcia danych ( art. 17 RODO )");
+            Console.WriteLine();
+            Console.WriteLine("d. prawo do żądania ograniczenia przetwarzania danych ( art. 18 RODO )");
+            Console.WriteLine();
+            Console.WriteLine("e. prawo do przenoszenia danych ( art. 20 RODO )");
+            Console.WriteLine();
+            Console.WriteLine("f. prawo do sprzeciwu wobec przetwarzania danych ( art. 21 RODO ).");
+            Console.WriteLine();
+            Console.WriteLine("Część uprawnień nie znajduje zastosowania w stosunku do danych osobowych przetwarzanych na podstawie art. 9 ust. 2 lit h) RODO, w tym w szczególności wobec danych przetwarzanych w ramach dokumentacji medycznej i innych przetwarzanych w oparciu o ww. przesłankę.");
+            Console.WriteLine();
+            Console.WriteLine("7. Osobie, której dane dotyczą ma prawo wniesienia skargi do organu nadzorczego właściwego dla ochrony danych osobowych.");
+            Console.WriteLine();
+            Console.WriteLine("8. podanie danych osobowych w rejestracji online jest dobrowolne, zaś osoba, której dane dotyczą może odmówić ich podania. Niepodanie danych będzie skutkować brakiem możliwości umówienia się na wizytę oraz konieczność skontaktowania się z Administratorem celem skutecznego umówienia wizyty i skorzystania ze świadczeń opieki zdrowotnej.");
+            Console.WriteLine();
+
+        }
+    }
+}
