--- Menu_template
@@ -1,31 +1,0 @@
-            
-            
-            /*
-            Console.Clear();
-            Console.WriteLine("------");
-            Console.WriteLine("");
-            Console.WriteLine("");
-            Console.WriteLine("");
-            Console.WriteLine("");
-            String[] opcje = { "------", "------", "------", "------" };
-            Menu_Config mainMenu = new Menu_Config(opcje);
-            int opt = mainMenu.Uruchom(opcje.Length);
-            switch (opt)
-            {
-                case 0:
-                
-                break;
-
-                case 1:
-                
-                break;
-
-                case 2:
-
-                break;
-
-                case 3:
-                
-                break;
-            }
-            */
