+++ Przychodnia_Gdynia.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31829.152
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Przychodnia_Gdynia", "Przychodnia_Gdynia.csproj", "{8E5B5DE1-0AA7-4A55-A4F9-A2E91A4A5550}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8E5B5DE1-0AA7-4A55-A4F9-A2E91A4A5550}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E5B5DE1-0AA7-4A55-A4F9-A2E91A4A5550}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E5B5DE1-0AA7-4A55-A4F9-A2E91A4A5550}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E5B5DE1-0AA7-4A55-A4F9-A2E91A4A5550}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {41206218-BC2E-4E3C-9800-4A7BC1C3F1AE}
+	EndGlobalSection
+EndGlobal
