+++ User/UserPanel.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Przychodnia_Gdynia
+{
+    class UserPanel
+    {
+        public static void menu()
+        {
+            Console.WriteLine();
+        }
+    }
+}
